+++ NEWS.md
@@ -1,5 +1,11 @@
 # kmerize 0.7.0
 
+## New Features
+
+* reading / writing kmer lists in parquet format
+
+## Stability Improvements
+
 * revisiting tests and adding CI
 
 # kmerize 0.6.0